+++ src/valency_anndata/preprocessing/__init__.py
@@ -3,6 +3,7 @@ from ._rebuild_vote_matrix import rebuild_vote_matrix
 from ._impute import impute
 from ._qc import calculate_qc_metrics
 from ._highly_variable_statements import highly_variable_statements
+from ._filter import filter_participants, filter_statements
 
 
 __all__ = [
@@ -10,6 +11,8 @@ __all__ = [
     "impute",
     "calculate_qc_metrics",
     "highly_variable_statements",
+    "filter_participants",
+    "filter_statements",
 
     # Simple re-export of scanpy.
     "neighbors",